requirements.txt
einops==0.8.0
joblib==1.3.2
matplotlib==3.5.3
numpy==1.21.2
pandas==2.2.2
PyYAML==6.0.1
scipy==1.7.3
torch==1.10.0
torchvision==0.11.0
tqdm==4.61.2
